Fix Issue #51 with DefaultHouseAds
Fixed two issues related to House Ads.

1. With a default house ad user control that has an extra period in it's
namespace such as: AdRotator.Examples.Windows8.1 Visual Studio decorates
the name by adding an underscore infront of this extra period. Our
reflection logic needs to account for this. When looking up the
AssemblyName we need to do our lookup without the added underscore, but
when resolving the classname we need to have the underscore. This
solution seems to work.

2. We were calling the async LoadCachedAd method from a catch block
which won't work because we can't await in a catch block